+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Added interface `WithDimension` to `PDimension` and `PCanvas`.
+- README.md: added usage example.
+
 ## 1.0.0
 
 - Initial version.